Day 11 part 2

+++ src/main/java/net/tylerday/AdventofCode/day11/Item.java
@@ -1,7 +1,7 @@
 package net.tylerday.AdventofCode.day11;
 
 public class Item {
-    int val;
+    long val;
 
     public Item(String val) {
         this.val = Integer.parseInt(val);
