+++ src/main/java/net/tylerday/AdventofCode/day11/MonkeyOperation.java
@@ -7,13 +7,13 @@ public class MonkeyOperation {
         this.descriptor = descriptor;
     }
 
-    public int apply(int oldVal) {
-        int val = 0;
+    public long apply(long oldVal) {
+        long val = 0;
         switch (descriptor){
             case ("new = old * 19") -> val =  oldVal * 19;
             case ("new = old + 6") -> val = oldVal + 6;
             case ("new = old * old") -> val = oldVal * oldVal;
-            case ("new = old + 3") -> val = oldVal +3;
+            case ("new = old + 3") -> val = oldVal + 3;
             case ("new = old * 3") -> val = oldVal * 3;
             case ("new = old + 7") -> val = oldVal + 7;
             case ("new = old + 5") -> val = oldVal + 5;
